Refactor CAN coordinator and state machine
