+++ app/src/main/kotlin/de/digitalService/useID/ui/screens/can/IdentificationCanPinForgotten.kt
@@ -52,7 +52,7 @@ interface IdentificationCanPinForgottenViewModelInterface {
 
 @HiltViewModel
 class IdentificationCanPinForgottenViewModel @Inject constructor(
-    val coordinator: CanCoordinator
+    private val coordinator: CanCoordinator
 ) : ViewModel(), IdentificationCanPinForgottenViewModelInterface {
 
     override fun onCancelIdentification() {
