+++ app/src/main/kotlin/de/digitalService/useID/ui/screens/can/SetupCanConfirmTransportPin.kt
@@ -67,7 +67,7 @@ class SetupCanConfirmTransportPinViewModel @Inject constructor(
     }
 
     override fun onConfirm() {
-        canCoordinator.confirmPinInput()
+        canCoordinator.onConfirmedPinInput()
     }
 
     override fun onReenterTransportPin() {
